Add new samples and refactoring
