+++ samples/security/tls/one_way_ssl/sample_k8s_artifacts/ballerinaKeystore.p12
