+++ samples/security/tls/one_way_ssl/sample_k8s_artifacts/kubernetes/sidecar-ingress.yaml
@@ -1,0 +1,23 @@
+---
+apiVersion: "extensions/v1beta1"
+kind: "Ingress"
+metadata:
+  annotations:
+    nginx.ingress.kubernetes.io/ssl-passthrough: "false"
+    kubernetes.io/ingress.class: "nginx"
+  finalizers: []
+  labels:
+    app: "tls"
+  name: "sidecar"
+  ownerReferences: []
+spec:
+  rules:
+  - host: "sidecar.com"
+    http:
+      paths:
+      - backend:
+          serviceName: "sidecar"
+          servicePort: 9090
+        path: "/"
+  tls:
+  - hosts: []
