+++ samples/security/tls/one_way_ssl/sample_k8s_artifacts/kubernetes/sidecar-svc.yaml
@@ -1,0 +1,20 @@
+---
+apiVersion: "v1"
+kind: "Service"
+metadata:
+  annotations: {}
+  finalizers: []
+  labels:
+    app: "tls"
+  name: "sidecar"
+  ownerReferences: []
+spec:
+  externalIPs: []
+  loadBalancerSourceRanges: []
+  ports:
+  - port: 9095
+    protocol: "TCP"
+    targetPort: 9095
+  selector:
+    app: "tls"
+  type: "NodePort"
