+++ samples/transactions/ballerina-springboot-transactions/hotel_service_springboot/src/main/java/ballerina/sidecar/samples/transactions/HotelReservationReq.java
@@ -1,0 +1,57 @@
+/*
+*  Copyright (c) 2017, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+*
+*  WSO2 Inc. licenses this file to you under the Apache License,
+*  Version 2.0 (the "License"); you may not use this file except
+*  in compliance with the License.
+*  You may obtain a copy of the License at
+*
+*    http://www.apache.org/licenses/LICENSE-2.0
+*
+*  Unless required by applicable law or agreed to in writing,
+*  software distributed under the License is distributed on an
+*  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+*  KIND, either express or implied.  See the License for the
+*  specific language governing permissions and limitations
+*  under the License.
+*/
+
+package ballerina.sidecar.samples.transactions;
+
+public class HotelReservationReq {
+    private final String fullName;
+    private final String checkIn;
+    private final String checkOut;
+    private final int rooms;
+
+    public HotelReservationReq() {
+        fullName = "";
+        checkIn = "";
+        checkOut = "";
+        rooms = 0;
+    }
+
+
+    public HotelReservationReq(String fullName, String checkIn, String checkOut, int rooms) {
+        this.fullName = fullName;
+        this.checkIn = checkIn;
+        this.checkOut = checkOut;
+        this.rooms = rooms;
+    }
+
+    public String getFullName() {
+        return fullName;
+    }
+
+    public String getCheckIn() {
+        return checkIn;
+    }
+
+    public String getCheckOut() {
+        return checkOut;
+    }
+
+    public int getRooms() {
+        return rooms;
+    }
+}
