+++ samples/transactions/ballerina-springboot-transactions/hotel_service_springboot/src/main/resources/schema.sql
