+++ samples/transactions/ballerina-springboot-transactions/hotel_service_springboot/src/test/hotel_reservation_req.json
