--- samples/transactions/ballerina-springboot-transactions/springboot-hotel-service/src/main/resources/schema.sql
@@ -1,2 +1,0 @@
-drop table BOOKINGS if exists;
-create table BOOKINGS(ID serial, FIRST_NAME varchar(10) NOT NULL);
